Gitlab-Backend-main/requirements.txt
fastapi
uvicorn
httpx
openai
python-dotenv
openai>=1.0.0
httpx
python-dotenv
